Refactored Parameters

+++ Remove-OMSWorkspaceStorageInsights.ps1
@@ -1,7 +1,24 @@
 ﻿param(
-$WorkspaceName,
-$StorageAccountName,
-$subscriptionName)
+[object]$WebhookData
+)
+
+if ($WebhookData -ne $null) {
+
+    $params = $WebhookData.RequestBody | Convert-From JSON
+
+    $subscriptionName = $params.subscriptionName
+    $WorkspaceName = $params.WorkspaceName
+    $StorageAccountName = $params.storageAccountName
+
+    Write-Output $subscriptionName
+    Write-Output $WorkspaceName
+    Write-Output $StorageAccountName
+
+}
+else {
+    Write-Output "Error, there was no HTTP Payload Data in the Request - exiting"
+    exit
+}
 
 #Add Service Principal Login
 $connectionName = "AzureRunAsConnection"
@@ -18,16 +35,7 @@ try
 
     "Logging in to Azure..."
 
-    Add-AzureRmAccount `
-
-        -ServicePrincipal `
-
-        -TenantId $servicePrincipalConnection.TenantId `
-
-        -ApplicationId $servicePrincipalConnection.ApplicationId `
-
-        -CertificateThumbprint $servicePrincipalConnection.CertificateThumbprint 
-
+    Add-AzureRmAccount -ServicePrincipal TenantId $servicePrincipalConnection.TenantId -ApplicationId $servicePrincipalConnection.ApplicationId -CertificateThumbprint $servicePrincipalConnection.CertificateThumbprint 
 }
 
 catch {
@@ -50,16 +58,39 @@ catch {
 
 }
 
-Set-AzureRmContext -Subscription $subscriptionName
+try {
+    Set-AzureRmContext -SubscriptionName $subscriptionName
+}
 
-$workspace = (Get-AzureRmOperationalInsightsWorkspace).Where({$_.Name -eq $WorkspaceName})
+catch {
+    if ($subscriptionName -eq $null) {
+        $ErrorMessage = "Parameter for the Subscription Name is null. Check your parameters"
+        Write-Error -Message $ErrorMessage
+        throw $ErrorMessage
+    }
+    else {
+        Write-Error -Message $_.Exception
+        throw $_.Exception
+    }
+}
 
-if ($workspace -ne $null)
+
+
+try {
+    $workspace = (Get-AzureRmOperationalInsightsWorkspace).Where({$_.Name -eq $WorkspaceName})    
+}
+
+catch {
+    Write-Error -Message $_.Exception
+    throw $_.Exception
+    }
+
+if ($workspace -ne $null -and $StorageAccountName -ne $null)
     {
     Remove-AzureRmOperationalInsightsStorageInsight -ResourceGroupName $workspace.ResourceGroupName -WorkspaceName $workspace.Name -Name $StorageAccountName -Force
     }
 else 
     {
-    write-error -Message "The OMS Workspace $($WorkspaceName) could not be found - Please make sure the Workspace Name is correct and you are targeting the correct subscription"
+    write-error -Message "The OMS Workspace $WorkspaceName, or the Storage Account $StorageAccountName could not be found - Please make sure the Workspace Name is correct and you are targeting the correct subscription"
     throw $_.Exception
     }
